Refactored recipe views into new files
